requirements.txt
torch
diffusers
transformers
accelerate
rembg
gradio
Pillow
numpy
scipy
onnxruntime-gpu
peft
